Struct rp_pac::pio::StateMachine

pub struct StateMachine(pub *mut u8);

Tuple Fields
0: *mut u8
Implementations

impl StateMachine

pub fn clkdiv(self) -> Reg<SmClkdiv, RW>

Clock divisor register for state machine 3 Frequency = clock freq / (CLKDIV_INT + CLKDIV_FRAC / 256)

pub fn execctrl(self) -> Reg<SmExecctrl, RW>

Execution/behavioural settings for state machine 3

pub fn shiftctrl(self) -> Reg<SmShiftctrl, RW>

Control behaviour of the input/output shift registers for state machine 3

pub fn addr(self) -> Reg<SmAddr, RW>

Current instruction address of state machine 3

pub fn instr(self) -> Reg<SmInstr, RW>

Read to see the instruction currently addressed by state machine 3’s program counter Write to execute an instruction immediately (including jumps) and then resume execution.

pub fn pinctrl(self) -> Reg<SmPinctrl, RW>

State machine pin control