Struct rp_pac::rosc::Rosc

pub struct Rosc(pub *mut u8);

Tuple Fields
0: *mut u8
Implementations

impl Rosc

pub fn ctrl(self) -> Reg<Ctrl, RW>

Ring Oscillator control

pub fn freqa(self) -> Reg<Freqa, RW>

The FREQA & FREQB registers control the frequency by controlling the drive strength of each stage The drive strength has 4 levels determined by the number of bits set Increasing the number of bits set increases the drive strength and increases the oscillation frequency 0 bits set is the default drive strength 1 bit set doubles the drive strength 2 bits set triples drive strength 3 bits set quadruples drive strength

pub fn freqb(self) -> Reg<Freqb, RW>

For a detailed description see freqa register

pub fn dormant(self) -> Reg<u32, RW>

Ring Oscillator pause control This is used to save power by pausing the ROSC On power-up this field is initialised to WAKE An invalid write will also select WAKE Warning: setup the irq before selecting dormant mode

pub fn div(self) -> Reg<Div, RW>

Controls the output divider

pub fn phase(self) -> Reg<Phase, RW>

Controls the phase shifted output

pub fn status(self) -> Reg<Status, RW>

Ring Oscillator Status

pub fn randombit(self) -> Reg<Randombit, RW>

This just reads the state of the oscillator output so randomness is compromised if the ring oscillator is stopped or run at a harmonic of the bus frequency

pub fn count(self) -> Reg<Count, RW>

A down counter running at the ROSC frequency which counts to zero and stops. To start the counter write a non-zero value. Can be used for short software pauses when setting up time sensitive hardware.