Struct rp_pac::sio::Interp

pub struct Interp(pub *mut u8);

Tuple Fields
0: *mut u8
Implementations

impl Interp

pub fn accum0(self) -> Reg<u32, RW>

Read/write access to accumulator 0

pub fn accum1(self) -> Reg<u32, RW>

Read/write access to accumulator 1

pub fn base0(self) -> Reg<u32, RW>

Read/write access to BASE0 register.

pub fn base1(self) -> Reg<u32, RW>

Read/write access to BASE1 register.

pub fn base2(self) -> Reg<u32, RW>

Read/write access to BASE2 register.

pub fn pop_lane0(self) -> Reg<u32, R>

Read LANE0 result, and simultaneously write lane results to both accumulators (POP).

pub fn pop_lane1(self) -> Reg<u32, R>

Read LANE1 result, and simultaneously write lane results to both accumulators (POP).

pub fn pop_full(self) -> Reg<u32, R>

Read FULL result, and simultaneously write lane results to both accumulators (POP).

pub fn peek_lane0(self) -> Reg<u32, R>

Read LANE0 result, without altering any internal state (PEEK).

pub fn peek_lane1(self) -> Reg<u32, R>

Read LANE1 result, without altering any internal state (PEEK).

pub fn peek_full(self) -> Reg<u32, R>

Read FULL result, without altering any internal state (PEEK).

pub fn ctrl_lane0(self) -> Reg<Interp1ctrlLane0, RW>

Control register for lane 0

pub fn ctrl_lane1(self) -> Reg<Interp1ctrlLane1, RW>

Control register for lane 1

pub fn accum0_add(self) -> Reg<Interp1accum0add, RW>

Values written here are atomically added to ACCUM0 Reading yields lane 0’s raw shift and mask value (BASE0 not added).

pub fn accum1_add(self) -> Reg<Interp1accum1add, RW>

Values written here are atomically added to ACCUM1 Reading yields lane 1’s raw shift and mask value (BASE1 not added).

pub fn base_1and0(self) -> Reg<u32, W>

On write, the lower 16 bits go to BASE0, upper bits to BASE1 simultaneously. Each half is sign-extended to 32 bits if that lane’s SIGNED flag is set.