Struct rp_pac::timer::Timer

pub struct Timer(pub *mut u8);

Controls time and alarms time is a 64 bit value indicating the time in usec since power-on timeh is the top 32 bits of time & timel is the bottom 32 bits to change time write to timelw before timehw to read time read from timelr before timehr An alarm is set by setting alarm_enable and writing to the corresponding alarm register When an alarm is pending, the corresponding alarm_running signal will be high An alarm can be cancelled before it has finished by clearing the alarm_enable When an alarm fires, the corresponding alarm_irq is set and alarm_running is cleared To clear the interrupt write a 1 to the corresponding alarm_irq

Tuple Fields
0: *mut u8
Implementations

impl Timer

pub fn timehw(self) -> Reg<u32, W>

Write to bits 63:32 of time always write timelw before timehw

pub fn timelw(self) -> Reg<u32, W>

Write to bits 31:0 of time writes do not get copied to time until timehw is written

pub fn timehr(self) -> Reg<u32, R>

Read from bits 63:32 of time always read timelr before timehr

pub fn timelr(self) -> Reg<u32, R>

Read from bits 31:0 of time

pub fn armed(self) -> Reg<Armed, RW>

Indicates the armed/disarmed status of each alarm. A write to the corresponding ALARMx register arms the alarm. Alarms automatically disarm upon firing, but writing ones here will disarm immediately without waiting to fire.

pub fn timerawh(self) -> Reg<u32, R>

Raw read from bits 63:32 of time (no side effects)

pub fn timerawl(self) -> Reg<u32, R>

Raw read from bits 31:0 of time (no side effects)

pub fn dbgpause(self) -> Reg<Dbgpause, RW>

Set bits high to enable pause when the corresponding debug ports are active

pub fn pause(self) -> Reg<Pause, RW>

Set high to pause the timer

pub fn intr(self) -> Reg<Int, RW>

Raw Interrupts

pub fn inte(self) -> Reg<Int, RW>

Interrupt Enable

pub fn intf(self) -> Reg<Int, RW>

Interrupt Force

pub fn ints(self) -> Reg<Int, RW>

Interrupt status after masking & forcing

pub fn alarm(self, n: usize) -> Reg<u32, RW>

Arm alarm 0, and configure the time it will fire. Once armed, the alarm fires when TIMER_ALARM0 == TIMELR. The alarm will disarm itself once it fires, and can be disarmed early using the ARMED status register.